backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.13.0
openai==2.31.0
python-dotenv==1.2.2
httpx==0.28.1
numpy==2.2.5
